Tests/LibWeb/Text/expected/XHR/XMLHttpRequest-responseURL.txt

responseURL before: ''
responseURL after: 'data:text/html,hello'
